// src/app/Model/BBSOrdertableModel.ts
import { bbsOrderShapeImageModel } from './bbsOrderShapeImageModel';

export interface BBSOrdertableModel {
  BBSID: number;
  BarCAB: boolean;
  BarEachQty: number;
  BarID: number;
  BarLength: string;
  BarMark: string;
  BarMemberQty: number;
  BarSTD: boolean;
  BarShapeCode: string;
  BarSize: string;
  BarSort: number;
  BarTotalQty: number;
  BarType: string;
  BarWeight: number;
  Cancelled: boolean;
  CustomerCode: string;
  ElementMark: string;
  JobID: number;
  PartGUID: string;
  PinSize: string;
  ProjectCode: string;
  Remarks: string;
  TeklaGUID: string;
  UpdateDate: string;
  shapeAutoCalcFormula1: string;
  shapeAutoCalcFormula2: string;
  shapeAutoCalcFormula3: string;
  shapeDefaultValue: string;
  shapeHeightCheck: string;
  shapeLengthFormula: string;
  shapeParType: string;
  shapeParaValidator: string;
  shapeParameters: string;
  shapeTransport: number;
  shapeTransportValidator: string;
  // shapeImageValue : bbsOrderShapeImageModel[]
  A: string;
  B: string;
  C: string;
  D: string;
  E: string;
  F: string;
  G: string;
  H: string;
  I: string;
  J: string;
  K: string;
  L: string;
  M: string;
  N: string;
  O: string;
  P: string;
  Q: string;
  R: string;
  S: string;
  T: string;
  U: string;
  V: string;
  W: string;
  X: string;
  Y: string;
  Z: string;
}